# docs/api_spec.yaml
openapi: 3.0.3
info:
  title: Distributed Synchronization System API
  description: |
    API untuk Distributed Synchronization System yang mengimplementasikan:
    - Distributed Lock Manager (Raft Consensus)
    - Distributed Queue System (Consistent Hashing)
    - Distributed Cache Coherence (MESI Protocol)
  version: 1.0.0
  contact:
    name: API Support
    email: support@distributed-sync.com
servers:
  - url: http://localhost:5000
    description: Node 1 (Development)
  - url: http://localhost:5001
    description: Node 2 (Development)
  - url: http://localhost:5002
    description: Node 3 (Development)
...
paths:
...
  /lock/acquire:
    post:
      tags:
        - Lock Manager
      summary: Acquire distributed lock
      description: Acquire a shared or exclusive lock on a resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LockAcquireRequest'
      responses:
        '200':
          description: Lock acquired successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  resource_id:
                    type: string
                  lock_type:
                    type: string
        '408':
          description: Lock acquisition timeout
        '409':
          description: Deadlock detected
        '503':
          description: Not leader, redirect to leader
  /lock/release:
    post:
      tags:
        - Lock Manager
      summary: Release distributed lock
      description: Release a previously acquired lock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LockReleaseRequest'
      responses:
        '200':
          description: Lock released successfully
        '404':
          description: Lock not found
        '503':
          description: Not leader
...
  /queue/produce:
    post:
      tags:
        - Queue
      summary: Produce message
      description: Produce a message to the distributed queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProduceRequest'
      responses:
        '200':
          description: Message produced
          content:
            application/json:
              schema:
                type: object
                properties:
                  message_id:
                    type: string
                  partition:
                    type: integer
...
  /cache/write:
    post:
      tags:
        - Cache
      summary: Write to cache
      description: Write a value to distributed cache
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CacheWriteRequest'
      responses:
        '200':
          description: Write successful
        '503':
          description: Not leader
...
components:
  schemas:
    LockAcquireRequest:
      type: object
      required:
        - resource_id
        - lock_type
        - client_id
      properties:
        resource_id:
          type: string
          description: Resource to lock
        lock_type:
          type: string
          enum: [shared, exclusive]
        client_id:
          type: string
          description: Client identifier
        timeout:
          type: number
          default: 30.0
          description: Timeout in seconds
    LockReleaseRequest:
      type: object
      required:
        - resource_id
        - client_id
      properties:
        resource_id:
          type: string
        client_id:
          type: string
...
    ProduceRequest:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          description: Message payload
        partition_key:
          type: string
          description: Key for partition selection
...
    CacheWriteRequest:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
        value:
          type: object
...